requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.2
djangorestframework==3.14.0
drf-yasg==1.21.7
factory-boy==3.3.0
Faker==23.1.0
inflection==0.5.1
MarkupSafe==2.1.5
packaging==23.2
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
setuptools==69.0.3
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.0
